get local address from peer (besides port), provide it back
This change allows the helper to accept both the old ``PORT`` and the new
``ADDRESS:PORT``, where ``ADDRESS`` is expected to be the local outgoing
interface address of the established connection as the peer sees it, i.e. not
as the helper sees it, since it may have been changed by NAT.

The new information is saved locally as ``PUB_ADDR:PORT/?l=LOC_ADDR`` if
there's not an identical entry in the file.  A similarly formatted entry may
be provided back to the peer.  When the peer builds the URL (even if it knows
nothing about local addresses), the URL will reflect both the public and local
addresses of the other peer, which should help telling apart connections not
working because of NAT (a normal case) or because of other reasons.

Old ``PUB_ADDR:PORT`` entries are kept and may also be returned to peers.