Include protocol in peer discovery
Only the details of a peer of the same protocol is sent back to the contacting
peer.  This allows us to use the same peer locator service to register peers
for different protocols.  The meaning of the protocol is left to the
probe (except for old probes, where plain HTTP is assumed).